% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova}
\alias{anova}
\title{A function to construct different types of ANOVA models}
\usage{
anova(
  data = NULL,
  type = c("between", "within", "mixed"),
  formula = NULL,
  outcome = NULL,
  between = NULL,
  within = NULL,
  id = NULL,
  detailed = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame containing the variables to be specified in the model.}

\item{type}{Unnecessary if formula is specified correctly: a character string specifying the type of ANOVA as "between" subjects, "within" subjects or (if factorial) "mixed."}

\item{formula}{An optional formula specifying the ANOVA model to be fit.}

\item{outcome}{If formula is unspecified, a numeric variable named as a character.}

\item{between}{If formula is unspecified, a between-subjects factor variable named as a character.}

\item{within}{If formula is unspecified, a within-subjects factor variable named as a character.}

\item{id}{If formula is unspecified, a factor variable named as a character which uniquely identifies each subject.}

\item{detailed}{If TRUE, returns additional information in the ANOVA table.}
}
\value{
A data frame
}
\description{
A function to construct different types of ANOVA models
}
\examples{
# One-Way Between-Subjects ANOVA
anova(data = ToothGrowth, formula = len ~ dose)
anova(data = ToothGrowth, type = "between", outcome = "len", between = "dose")
# Two-Way Between-Subjects ANOVA
mtcars$am <- factor(mtcars$am)
anova(data = mtcars, formula = disp ~ vs*am)
anova(data = mtcars, type = "between", outcome = "disp", between = c("vs", "am"))
# One-Way Within-Subjects ANOVA
wrestlers <- PASWR2::HSWRESTLER \%>\% pivot_longer(cols = c("hwfat", "tanfat", "skfat"), names_to = "type", values_to = "measurement")
wrestlers$id <- factor(sort(rep(1:78, 3)))
anova(data = wrestlers, formula = measurement ~ type + Error(id/type))
anova(data = wrestlers, type = "within", outcome = "measurement", within = "type", id = "id")
# Two-Way Within-Subjects ANOVA
expt <- data.frame(PID = factor(sort(rep(1:50,4))), image = factor(rep(c("Angry","Angry","Happy","Happy"),50)), music = rep(c("Disney","Horror"),100), stress = rnorm(200,60,10))
anova(data = expt, formula = stress ~ image*music + Error(PID/(image*music)))
anova(data = expt, type = "within", outcome = "stress", within = c("image", "music"), id = "PID")
# Two-Way Mixed ANOVA
ChickWeight$Time <- factor(ChickWeight$Time)
ChickWeight$Chick <- factor(ChickWeight$Chick)
anova(data = ChickWeight, formula = weight ~ Diet*Time + Error(Chick/Time))
anova(data = ChickWeight, type = "mixed", outcome = "weight", within = "Time", between = "Diet", id = "Chick")
}
