% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{regress}
\alias{regress}
\title{A function to construct different kinds of linear regression models}
\usage{
regress(formula = NULL, data = NULL, ...)
}
\arguments{
\item{formula}{A symbolic description of the model to be fitted of the form y ~ x...}

\item{data}{A data frame containing the variables specified in the formula and contained in the model.}
}
\value{
An object of \code{class} \code{"lm"}.
}
\description{
A function to construct different kinds of linear regression models
}
\examples{
regress(formula = disp~am, data = mtcars)
}
