% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_calcs_regression.R
\name{regression.calcs}
\alias{regression.calcs}
\title{The regression model method for the \code{calcs} function from the \code{doingRbest} package}
\usage{
regression.calcs(model = NULL, predictors = NULL, outcome = NULL)
}
\arguments{
\item{model}{The stored model object created from the \code{regress} function in the \code{doingRbest} package.}

\item{predictors}{An optional character vector of custom names for the x-variables in the model.}

\item{outcome}{An optional custom name (character vector of length 1) for the y-variable in the model.}

\item{...}{Extra named arguments that may be available upon package updates.}
}
\value{
An HTML page with a series of solved equations (in a separate window or the Viewer pane)
}
\description{
The regression model method for the \code{calcs} function from the \code{doingRbest} package
}
\examples{
# Simple Regression Calculations
model <- regress(disp ~ mpg, data = mtcars)
calcs(model)
# Multiple Regression Calculations (limited with >2 predictors)
model <- regress(disp ~ mpg + wt, data = mtcars)
calcs(model)
# Moderated Regression Calculations 
model <- regress(disp ~ mpg*wt, data = mtcars)
calcs(model)
}
