% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_graph_regression.R
\name{regression.graph}
\alias{regression.graph}
\title{The regression model method for the \code{graph} function from the \code{doingRbest} package}
\usage{
regression.graph(model, predictors = c(), outcome = NULL, ...)
}
\arguments{
\item{model}{The stored model object created from the \code{regress} function in the \code{doingRbest} package.}

\item{predictors}{An optional character vector of custom names for the x-variables in the model.}

\item{outcome}{An optional custom name (character vector of length 1) for the y-variable in the model.}

\item{...}{Extra named arguments that may be available upon package updates.}
}
\value{
A \code{plotly} plot (in a separate window or the Viewer pane)
}
\description{
The regression model method for the \code{graph} function from the \code{doingRbest} package
}
\examples{
# Simple Regression Graph (scatterplot with line)
model <- regress(disp ~ mpg, data = mtcars)
graph(model)
# Multiple Regression Graph (3-D scatterplot with plane - 2 predictors maximum to render in 3-D)
model <- regress(disp ~ mpg + wt, data = mtcars)
graph(model)
# Moderated Regression Graph (3-D scatterplot with curved plane)
model <- regress(disp ~ mpg*wt, data = mtcars)
graph(model)
# Moderated Regression Simple Slopes Graph
model <- regress(disp ~ mpg*wt, data = mtcars)
graph(model, sim_slopes = TRUE)
}
