% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_hyp_regression.R
\name{regression.hyp}
\alias{regression.hyp}
\title{The regression model method for the \code{hyp} function from the \code{doingRbest} package}
\usage{
regression.hyp(model = NULL, ...)
}
\arguments{
\item{model}{The stored model object created from the \code{regress} function in the \code{doingRbest} package.}
}
\value{
An HTML page with a series of null and alternative hypotheses (in a separate window or the Viewer pane)
}
\description{
The regression model method for the \code{hyp} function from the \code{doingRbest} package
}
\examples{
# Simple Regression Hypothesis Test
model <- regress(disp ~ mpg, data = mtcars)
hyp(model)
# Multiple Regression Hypothesis Tests
model <- regress(disp ~ mpg + wt, data = mtcars)
hyp(model)
# Moderated Regression Hypothesis Tests
model <- regress(disp ~ mpg*wt, data = mtcars)
hyp(model)
}
