% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_output_regression.R
\name{regression.output}
\alias{regression.output}
\title{The regression model method for the \code{output} function from the \code{doingRbest} package}
\usage{
regression.output(
  model = NULL,
  predictors = NULL,
  outcome = NULL,
  interpretation = FALSE,
  ...
)
}
\arguments{
\item{model}{The stored model object created from the \code{regress} function in the \code{doingRbest} package.}

\item{predictors}{An optional character vector of custom names for the x-variables in the model.}

\item{outcome}{An optional custom name (character vector of length 1) for the y-variable in the model.}

\item{interpretation}{Set equal to \code{TRUE} for an explanatory paragraph beneath the output table interpreting the numeric values.}

\item{...}{Extra named arguments that may be available upon package updates.}

\item{sim_slopes}{For moderated regression output, set equal to \code{TRUE} for a simple slopes analysis in place of a traditional output table for the analysis.}

\item{flip_moderator}{For moderated regression output when \code{sim_slopes = TRUE}, alternate the position between predictor and moderator variables in the output.}
}
\value{
An output table as an HTML object (in a separate window or the Viewer pane)
}
\description{
The regression model method for the \code{output} function from the \code{doingRbest} package
}
\examples{
# Simple Regression Output Table
model <- regress(disp ~ mpg, data = mtcars)
output(model)
# Multiple Regression Output Table
model <- regress(disp ~ mpg + wt + hp, data = mtcars)
output(model)
# Moderated Regression Output Table
model <- regress(disp ~ mpg*wt, data = mtcars)
output(model)
# Moderated Regression Simple Slopes Output Table
model <- regress(disp ~ mpg*wt, data = mtcars)
output(model, sim_slopes = TRUE)
}
