% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_sampdist_regression.R
\name{regression.sampdist}
\alias{regression.sampdist}
\title{The regression model method for the \code{sampdist} function from the \code{doingRbest} package}
\usage{
regression.sampdist(
  model = NULL,
  Fdist = FALSE,
  tdist = FALSE,
  slope = NULL,
  predictors = NULL,
  outcome = NULL,
  ...
)
}
\arguments{
\item{model}{The stored model object created from the \code{regress} function in the \code{doingRbest} package to run a statistical model.}

\item{Fdist}{For regression models, set equal to \code{TRUE} to display a sampling distribution for the whole model F-statistic.}

\item{tdist}{For regression models, set equal to \code{TRUE} to display a sampling distribution for the first slope coefficient's t-statistic (if not named in \code{slope} argument).}

\item{predictors}{An optional character vector of custom names for the x-variables in the model.}

\item{outcome}{An optional custom name (character vector of length 1) for the y-variable in the model.}

\item{...}{Extra named arguments that may be available upon package updates.}
}
\value{
A \code{plotly} plot
}
\description{
The regression model method for the \code{sampdist} function from the \code{doingRbest} package
}
\examples{
# Simple Regression Sampling Distribution(s)
model <- regress(disp ~ mpg, data = mtcars)
sampdist(model)
# Multiple Regression Sampling Distribution(s)
model <- regress(disp ~ mpg + wt, data = mtcars)
sampdist(model)
# Moderated Regression Sampling Distribution(s)
model <- regress(disp ~ mpg*wt, data = mtcars)
sampdist(model)
}
