% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_hyp_ttest.R
\name{ttest.hyp}
\alias{ttest.hyp}
\title{The ttest model method for the \code{hyp} function from the \code{doingRbest} package}
\usage{
ttest.hyp(model = NULL, ...)
}
\arguments{
\item{model}{The stored model object created from the \code{ttest} function in the \code{doingRbest} package.}
}
\value{
An HTML page with a series of null and alternative hypotheses (in a separate window or the Viewer pane)
}
\description{
The ttest model method for the \code{hyp} function from the \code{doingRbest} package
}
\examples{
# Single Sample t-test Hypothesis Test
model <- ttest(outcome = "disp", data = mtcars, type = "s", mu = 100)
hyp(model, outcome = "Fuel Displacement")
# Independent Samples t-test Hypothesis Test
model <- ttest(outcome = "hardness", predictor = "location", data = PASWR2::APPLE, type = "independent")
hyp(model)
# Dependent Samples t-test Hypothesis Test for "wide" format data
early <- c(200.1, 190.9, 192.7, 213, 241.4, 196.9, 172.2, 185.5, 205.2, 193.7)
late <- c(392.9, 393.2, 345.1, 393, 434, 427.9, 422, 383.9, 392.3, 352.2)
my_data <- data.frame(early, late)
model <- ttest(outcome = "weight", predictor = "group", data = my_data, data_format = "wide", type = "dependent", cols = c("early", "late"))
hyp(model)
# Dependent Samples t-test Hypothesis Test for "long" format data
SubID <- factor(sort(rep(1:10,2)))
Grp <- factor(rep(c("Ctrl","Trt"),10))
DV <- rnorm(20,10,2)
the_data <- data.frame(SubID, Grp, DV)
model <- ttest(outcome = "DV", predictor = "Grp", data = the_data, data_format = "long", type = "dependent")
hyp(model)
}
