% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_output_regression.R
\name{regression.output}
\alias{regression.output}
\title{A generic function used to create a table of analysis-specific output for common statistical models}
\usage{
regression.output(
  model = NULL,
  predictors = NULL,
  outcome = NULL,
  interpretation = FALSE,
  ...
)
}
\arguments{
\item{model}{The stored model object created from a function in the doingRbest package to run a statistical model.}

\item{predictors}{An optional character vector of custom names for the x-variables in the model.}

\item{outcome}{An optional custom name (character vector of length 1) for the y-variable in the model.}

\item{interpretation}{Set equal to \code{TRUE} for an explanatory paragraph beneath the output table interpreting the numeric values.}

\item{...}{Extra named arguments that may be available upon package updates.}

\item{sim_slopes}{For moderated regression or factorial ANOVA output, set equal to \code{TRUE} for a simple slopes/simple effects analysis in place of a traditional output table for the analysis.}
}
\value{
An output table as an HTML object (in a separate window or the Viewer pane)
}
\description{
A generic function used to create a table of analysis-specific output for common statistical models
}
\examples{
# Simple Regression Output Table
model <- regress(disp ~ mpg, data = mtcars)
output(model)
}
