% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_sampdist.R
\name{sampdist}
\alias{sampdist}
\title{A generic function used to generate a set of sampling distribution curves to compare test statistics to critical values for a given type of analysis}
\usage{
sampdist(
  model = NULL,
  predictors = NULL,
  outcome = NULL,
  Fdist = TRUE,
  tdist = TRUE,
  ...
)
}
\arguments{
\item{model}{The stored model object created from a function in the \code{doingRbest} package to run a statistical model.}

\item{predictors}{An optional character vector of custom names for the x-variables in the model.}

\item{outcome}{An optional custom name (character vector of length 1) for the y-variable in the model.}

\item{Fdist}{For regression models, set equal to \code{TRUE} to display a sampling distribution for the whole model F-statistic.}

\item{tdist}{For regression models, set equal to \code{TRUE} to display a sampling distribution for the first slope coefficient's t-statistic (if not named in \code{slope} argument).}

\item{...}{Extra named arguments that may be available upon package updates.}
}
\value{
A \code{plotly} plot
}
\description{
A generic function used to generate a set of sampling distribution curves to compare test statistics to critical values for a given type of analysis
}
\examples{
 Simple Regression Sampling Distribution(s)
model <- regress(disp ~ mpg, data = mtcars)
sampdist(model)
# Multiple Regression Sampling Distribution(s)
model <- regress(disp ~ mpg + wt, data = mtcars)
sampdist(model)
# Moderated Regression Sampling Distribution(s)
model <- regress(disp ~ mpg*wt, data = mtcars)
sampdist(model)
}
